+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Enh #1: Initial commit (@terabytesoftw)
 - Enh #2: Add dropdown cookbook selector `language` and toggle cookbok `selectorLanguage` classes (@terabytesoftw)
+- Enh #3: Add toggle cookbook `selectorTheme` class (@terabytesoftw)